+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
  - Add way to get users for workshops
+ - Added functionality search beneficiaries
 
 ## [0.1.0] - 15/11/2025
 ### Fixed